Module Dhcp_server

A DHCP server is composed of two sub-modules: Config and Input. The former deals with building a suitable configuration for using with the later.

v1.5.0 - homepage

DHCP Server Configuration

module Config : sig ... end

DHCP Leases (bindings)

module Lease : sig ... end

DHCP Input Packet Logic

module Input : sig ... end

The logic for handling a DHCP input packet is pure, the module does not perform any IO, it only returns a possible reply packet or event to be logged.